Use Phoenix.Param protocol to use slugs
Phoenix.Param protocol allows to define how we want to display urls when
using the url helpers. It also has a neat `@derive` module attribute
that is shorthand if all you want to do is use a different value in the
schema instead of the id. So we use that to make the `slug` the default
when doing something like `url_path(@conn, :action, @announcement)`.

It turns out we were explicitly passing `@announcement.id` in many
places, which removes the use of the protocol. So we update those places
to make use of the protocol.